moved main class selector to serial ui
